Completed validations and unit tests

+++ services/productCatalogue/src/main/java/org/bazaar/productCatalogue/sale/dto/SaleProduct.java
@@ -1,10 +1,12 @@
 package org.bazaar.productCatalogue.sale.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 
 @Builder
 public record SaleProduct(String name,
-                Double originalPrice,
+                BigDecimal originalPrice,
                 String category) {
 
 }
